ELK/requirements.txt
black==25.1.0
certifi==2025.1.31
click==8.1.8
elastic-transport==8.17.0
elasticsearch==8.17.1
mypy-extensions==1.0.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
six==1.17.0
tzdata==2025.1
urllib3==2.3.0
